Separate tabs into separate components
